last update database

+++ backend/migrations/20240815135938_admin.js
@@ -4,7 +4,7 @@
  */
 exports.up = function (knex) {
   return knex.schema.createTable("ADMIN", function (table) {
-    table.bigIncrements("ADMIN_ID"); // BIGINT IDENTITY column for admin ID with auto-increment
+    table.increments("ADMIN_ID"); // BIGINT IDENTITY column for admin ID with auto-increment
     table.string("NAME", 255).notNullable(); // VARCHAR(255) for name
     table.string("EMAIL", 255).notNullable(); // VARCHAR(255) for email
     table.string("PASSWORD", 255).notNullable(); // VARCHAR(255) for password
